move simulation functions to utils
